backend/requirements.txt
fastapi==0.115.12
pandas==2.3.0
pydantic==2.11.7
python-dotenv==1.1.0
python_dateutil==2.8.2
yfinance==0.2.62
uvicorn==0.34.0
numpy==1.26.4
requests==2.31.0
beautifulsoup4==4.12.2
plaid-python==16.0.0
supabase==2.16.0
openai==1.54.3
